!! TODO: Review this commit !!
1. Bumped babel-core@^7.0.0-bridge.0 to renamed package @babel/core@^7.6.4.

2. Switched from @babel/runtime-corejs2@^7.0.0 > renamed package babel-loader@^8.0.6 (this is not the corejs2 version, I have not tested for issues).

3. Added latest @babel/preset-env, and added required (by our React version, I believe) @babel/plugin-proposal-class-properties@^7.5.5

Rationale: I was getting TypeError when using some array methods; such as map. I believed this may have been caused by use of Babel6 instead of Babel7.
However, I am not sure that this was the case, so this commit may be removed before this branch is merged, if it causes tests to fail..